Tests/LibWeb/Text/expected/css/animated-background-image-timer-stops-after-layout-node-replacement.txt

active animations before replacement: 1
active animations after replacement: 1
active animations after hiding: 0
